+++ frontend/src/routes/group/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	import '../../app.css';
+	
+	let { children } = $props();
+</script>
+
+{@render children()}
